refactor(llm): hoist KV-slot value types into KvSlotTypes.hpp (KvSlotManager stage 1)
Move the pure-data multi-slot KV types verbatim out of LLM::Impl into
src/runner/KvSlotTypes.hpp: LinearStateSnapshot, KvSlotLocation, KvCacheSlot,
kSlotReuseMinPrefix. Impl keeps the *active* member instances (kv_slots_,
kv_active_slot_idx_, linear_state_snapshots_, ...). No behavior change; enables
the KV-slot selection policy + manager (and their unit tests) to live in their
own translation units in stages 2-3.

Verified: LLM.cpp compiles with real build_x86 (AXCL) flags on the bridge
(EXIT 0); braces balanced.
